# BE/docs/swagger.yaml
swagger: "2.0"
info:
  title: IT4409 Jira-Clone API
  description: |
    Backend API cho hệ thống quản lý dự án kiểu Jira.
    Clean Architecture (Go) + PostgreSQL (Supabase).
    
    **Hướng dẫn xác thực:**
    1. Gọi `/api/auth/register` hoặc `/api/auth/login` để lấy JWT token.
    2. Nhấn nút "Authorize" ở góc trên, nhập: `Bearer <token>`.
    3. Tất cả API có khóa 🔒 đều yêu cầu JWT.
  version: "2.0"
  contact:
    name: IT4409 Team
basePath: /
schemes:
  - http
securityDefinitions:
  BearerAuth:
    description: "Nhập JWT token. Ví dụ: Bearer eyJhbGciOiJI..."
    type: apiKey
    name: Authorization
    in: header

# ════════════════════════════════════════════════
#  DEFINITIONS (Cấu trúc dữ liệu)
# ════════════════════════════════════════════════
definitions:
  Envelope:
    type: object
    properties:
      status:
        type: integer
      message:
        type: string
      data: {}
...
  Board:
    type: object
    properties:
      id:
        type: string
      projectId:
        type: string
      name:
        type: string
      isDefault:
        type: boolean
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time
  BoardColumn:
    type: object
    properties:
      id:
        type: string
      boardId:
        type: string
      name:
        type: string
      position:
        type: integer
      statusMap:
        type: string
        enum: [todo, in_progress, in_review, done]
      createdAt:
        type: string
        format: date-time
  BoardDetail:
    type: object
    properties:
      board:
        $ref: "#/definitions/Board"
      columns:
        type: array
        items:
          $ref: "#/definitions/BoardColumn"
...
  Comment:
    type: object
    properties:
      id:
        type: string
      issueId:
        type: string
      userId:
        type: string
      content:
        type: string
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time
...
paths:
...
  /api/projects/{projectID}/boards:
    post:
      tags: [Boards]
      summary: Tạo bảng Kanban mới
      description: Tự động tạo 4 cột mặc định (To Do, In Progress, In Review, Done). Yêu cầu quyền member+.
      security: [{BearerAuth: []}]
      consumes: [application/json]
      parameters:
        - in: path
          name: projectID
          type: string
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [name]
            properties:
              name:
                type: string
                example: Sprint Board
      responses:
        201:
          description: Tạo bảng thành công
          schema:
            $ref: "#/definitions/Board"
    get:
      tags: [Boards]
      summary: Liệt kê bảng trong dự án
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: projectID
          type: string
          required: true
      responses:
        200:
          description: Danh sách bảng
          schema:
            type: array
            items:
              $ref: "#/definitions/Board"
  /api/boards/{boardID}:
    get:
      tags: [Boards]
      summary: Chi tiết bảng + danh sách cột
      description: Trả về bảng kèm tất cả cột, dùng để render giao diện Kanban.
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: boardID
          type: string
          required: true
      responses:
        200:
          description: Bảng + cột
          schema:
            $ref: "#/definitions/BoardDetail"
    patch:
      tags: [Boards]
      summary: Cập nhật bảng
      description: "Yêu cầu quyền: admin"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: boardID
          type: string
          required: true
        - in: body
          name: body
          schema:
            type: object
            properties:
              name:
                type: string
      responses:
        200:
          description: Cập nhật thành công
    delete:
      tags: [Boards]
      summary: Xóa bảng
      description: "Yêu cầu quyền: admin"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: boardID
          type: string
          required: true
      responses:
        204:
          description: Xóa thành công

  # ── COLUMNS (Người B) ──────────────────────────
  /api/boards/{boardID}/columns:
    post:
      tags: [Board Columns]
      summary: Thêm cột vào bảng
      description: "Yêu cầu quyền: admin"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: boardID
          type: string
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [name, statusMap]
            properties:
              name:
                type: string
                example: Testing
              position:
                type: integer
                example: 3
              statusMap:
                type: string
                enum: [todo, in_progress, in_review, done]
      responses:
        201:
          description: Thêm cột thành công
          schema:
            $ref: "#/definitions/BoardColumn"
  /api/boards/{boardID}/columns/{columnID}:
    patch:
      tags: [Board Columns]
      summary: Sửa tên/trạng thái cột
      description: "Yêu cầu quyền: admin"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: boardID
          type: string
          required: true
        - in: path
          name: columnID
          type: string
          required: true
        - in: body
          name: body
          schema:
            type: object
            properties:
              name:
                type: string
              statusMap:
                type: string
                enum: [todo, in_progress, in_review, done]
      responses:
        200:
          description: Sửa cột thành công
    delete:
      tags: [Board Columns]
      summary: Xóa cột
      description: "Yêu cầu quyền: admin"
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: boardID
          type: string
          required: true
        - in: path
          name: columnID
          type: string
          required: true
      responses:
        204:
          description: Xóa cột thành công
  /api/boards/{boardID}/columns/reorder:
    put:
      tags: [Board Columns]
      summary: Sắp xếp lại thứ tự cột (kéo thả)
      description: Gửi mảng columnIds theo thứ tự mới. Yêu cầu quyền member+.
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: boardID
          type: string
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              columnIds:
                type: array
                items:
                  type: string
                example: ["col-uuid-3", "col-uuid-1", "col-uuid-2", "col-uuid-4"]
      responses:
        200:
          description: Sắp xếp lại thành công

  # ── LABELS (Người B) ───────────────────────────
...
  /api/issues/{issueKey}/comments:
    post:
      tags: [Comments]
      summary: Thêm bình luận
      security: [{BearerAuth: []}]
      consumes: [application/json]
      parameters:
        - {in: path, name: issueKey, type: string, required: true}
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              content: {type: string, example: "Đã xem xong, cần sửa thêm."}
      responses:
        201: {description: Thêm thành công, schema: {$ref: "#/definitions/Comment"}}
    get:
      tags: [Comments]
      summary: Liệt kê bình luận
      security: [{BearerAuth: []}]
      parameters:
        - {in: path, name: issueKey, type: string, required: true}
        - {in: query, name: page, type: integer, default: 1}
        - {in: query, name: per_page, type: integer, default: 20}
      responses:
        200: {description: Danh sách bình luận}
  /api/comments/{commentID}:
    patch:
      tags: [Comments]
      summary: Sửa bình luận (chỉ tác giả)
      security: [{BearerAuth: []}]
      consumes: [application/json]
      parameters:
        - {in: path, name: commentID, type: string, required: true}
        - in: body
          name: body
          schema:
            type: object
            properties:
              content: {type: string}
      responses:
        200: {description: Sửa thành công}
    delete:
      tags: [Comments]
      summary: Xóa bình luận (tác giả hoặc admin)
      security: [{BearerAuth: []}]
      parameters:
        - {in: path, name: commentID, type: string, required: true}
      responses:
        200: {description: Xóa thành công}

  # ── ATTACHMENTS ────────────────────────────────
...
  /api/issues/{issueKey}/activity:
    get:
      tags: [Activity]
      summary: Lịch sử thay đổi của issue
      security: [{BearerAuth: []}]
      produces: [application/json]
      parameters:
        - {in: path, name: issueKey, type: string, required: true, description: "Issue key (e.g. PROJ-1)"}
        - {in: query, name: page, type: integer, description: "Trang (mặc định 1)"}
        - {in: query, name: per_page, type: integer, description: "Số item/trang (mặc định 20)"}
      responses:
        200:
          description: Danh sách activity
          schema:
            $ref: "#/definitions/Envelope"
  /api/projects/{projectID}/activity:
    get:
      tags: [Activity]
      summary: Luồng hoạt động dự án
      security: [{BearerAuth: []}]
      produces: [application/json]
      parameters:
        - {in: path, name: projectID, type: string, required: true}
        - {in: query, name: page, type: integer, description: "Trang (mặc định 1)"}
        - {in: query, name: per_page, type: integer, description: "Số item/trang (mặc định 20)"}
      responses:
        200:
          description: Danh sách activity của toàn project
          schema:
            $ref: "#/definitions/Envelope"

  # ── SEARCH ─────────────────────────────────────
...